% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_priors}
\alias{get_priors}
\title{Get Default Prior Values for Disaggregation Model}
\usage{
get_priors(data)
}
\arguments{
\item{data}{A \code{disag_data_mmap} object (output from \code{prepare_data_mmap}).}
}
\value{
A named list of prior specifications.
}
\description{
Calculates the default Penalized Complexity (PC) prior parameters and Gaussian
priors that will be used by \code{disag_model_mmap()} if the user does not
provide overrides.
}
\details{
The default priors are dynamic and depend on the input data:
\itemize{
  \item \strong{Range (Rho):} The lower bound \code{prior_rho_min} is set to
  1/3 of the diagonal length of the study area's bounding box.
  \item \strong{Spatial SD (Sigma):} The upper bound \code{prior_sigma_max}
  is set to the coefficient of variation of the polygon response counts.
}
}
\examples{
# Create minimal polygon and covariate inputs for one time point.
polygons <- sf::st_sf(
  area_id = 1:2,
  response = c(10, 12),
  geometry = sf::st_sfc(
    sf::st_polygon(list(rbind(c(0, 0), c(1, 0), c(1, 2), c(0, 2), c(0, 0)))),
    sf::st_polygon(list(rbind(c(1, 0), c(2, 0), c(2, 2), c(1, 2), c(1, 0)))),
    crs = 3857
  )
)
covariate <- terra::rast(
  ncols = 2, nrows = 2, xmin = 0, xmax = 2, ymin = 0, ymax = 2,
  crs = "EPSG:3857"
)
terra::values(covariate) <- c(1, 2, 3, 4)

data <- suppressMessages(prepare_data_mmap(
  polygon_shapefile_list = list(polygons),
  covariate_rasters_list = list(covariate),
  make_mesh = FALSE
))

# Inspect defaults and modify a prior for a later model fit.
defaults <- get_priors(data)
defaults[c("prior_rho_min", "prior_sigma_max")]
my_priors <- defaults
my_priors$prior_rho_prob <- 0.05
}
\keyword{external}
